requirements.txt
chromadb
fastapi
google-genai
langchain
langchain-chroma
langchain-community
langchain-google-vertexai
loguru
prometheus-fastapi-instrumentator
pydantic
pypdf
python-dotenv
tqdm
uvicorn